+++ StdAfx.h
@@ -31,7 +31,7 @@ template<typename T> inline void memzero(T & obj)
 
 #define countof(a) (sizeof (a) / sizeof (a[0]))
 
-static size_t MultiSzLen(LPCSTR src)
+static unsigned MultiSzLen(LPCSTR src)
 {
 	size_t len = 0;
 	size_t len1;
@@ -39,10 +39,10 @@ static size_t MultiSzLen(LPCSTR src)
 	{
 		len += len1 + 1;
 	}
-	return len;
+	return (unsigned)len;
 }
 
-static size_t MultiSzLen(LPCWSTR src)
+static unsigned MultiSzLen(LPCWSTR src)
 {
 	size_t len = 0;
 	size_t len1;
@@ -50,7 +50,7 @@ static size_t MultiSzLen(LPCWSTR src)
 	{
 		len += len1 + 1;
 	}
-	return len;
+	return (unsigned)len;
 }
 
 inline void AssignMultiSz(CStringW & dst, LPCSTR src)
